+++ CHANGELOG.md
@@ -61,6 +61,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   diff formatters. Built-in tools that do not function well as diff editing
   tools or as diff formatters will now report an error when used as such.
 
+* `jj diffedit` now accepts filesets to edit only the specified paths.
+
 ### Fixed bugs
 
 ### Packaging changes